sdk/tests/conformance/state/gl-samples-framebuffer-switch.html

<!--

/*
** Copyright (c) 2014 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL getParameter(SAMPLES) after framebuffer binding in WebGL 1.0</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>
<script>
"use strict";
description();

var wtu = WebGLTestUtils;
var gl = wtu.create3DContext('canvas', {antialias: true});
if (!gl)
    testFailed('context does not exist');
else {
    testPassed('context created');

    var default_samples = gl.getParameter(gl.SAMPLES);
    debug('Sample count while default framebuffer is bound: ' + samples);
    var fbo = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fbo);
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, 'no errors from setting up a framebuffer object');
    // The value of gl.SAMPLES is specified as implementation-dependent in GLES 2.0 spec section 3.2 Multisampling.
    // It should not change when framebuffer binding is changed, unlike in later versions of the GLES/WebGL spec.
    shouldBe('gl.getParameter(gl.SAMPLES)', 'default_samples');
}

var successfullyParsed = true;
finishTest();

</script>
</body>
</html>
